% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_to_kg.R
\name{lb_to_kg}
\alias{lb_to_kg}
\title{Pounds to Kilograms}
\usage{
lb_to_kg(lb)
}
\arguments{
\item{lb}{The argument for this function is lb. Lb is the symbol/abbreviation
for pounds. This is a numeric argument. The inputs must be greater than 0.02.}
}
\value{
The function returns a number, kg, which is the converted inputted
pounds into kilograms. The output is limited to two decimal
places for conciseness.
}
\description{
This function will take an input, in pounds, and return the
conversion to kilograms rounded to 2 decimal points for conciseness. Only
inputs greater than 0.02 can be inputted. If a smaller input is made, the
function will specify that it can't compute the conversion.
}
\examples{
lb_to_kg(10)
lb_to_kg(5)
}
